Expose custom repo methods on both wrapper and local implementations.
