+++ apps/explorer/lib/explorer/repo.ex
@@ -3,6 +3,8 @@ defmodule Explorer.Repo.Local do
     otp_app: :explorer,
     adapter: Ecto.Adapters.Postgres
 
+  use Explorer.Repo.RepoHelper
+
   require Logger
 
   alias Explorer.Repo.ConfigHelper
@@ -58,100 +60,7 @@ defmodule Explorer.Repo do
   use Fly.Repo, local_repo: Explorer.Repo.Local
   require Logger
 
+  use Explorer.Repo.RepoHelper
+
   def replica, do: __MODULE__
-
-  def logged_transaction(fun_or_multi, opts \\ []) do
-    transaction_id = :erlang.unique_integer([:positive])
-
-    Explorer.Logger.metadata(
-      fn ->
-        {microseconds, value} = :timer.tc(__MODULE__, :transaction, [fun_or_multi, opts])
-
-        milliseconds = div(microseconds, 100) / 10.0
-        Logger.debug(["transaction_time=", :io_lib_format.fwrite_g(milliseconds), ?m, ?s])
-
-        value
-      end,
-      transaction_id: transaction_id
-    )
-  end
-
-  @doc """
-  Chunks elements into multiple `insert_all`'s to avoid DB driver param limits.
-
-  *Note:* Should always be run within a transaction as multiple inserts may occur.
-  """
-  def safe_insert_all(kind, elements, opts) do
-    returning = opts[:returning]
-
-    elements
-    |> Enum.chunk_every(500)
-    |> Enum.reduce({0, []}, fn chunk, {total_count, acc} ->
-      {count, inserted} =
-        try do
-          insert_all(kind, chunk, opts)
-        rescue
-          exception ->
-            old_truncate = Application.get_env(:logger, :truncate)
-            Logger.configure(truncate: :infinity)
-
-            Logger.error(fn ->
-              [
-                "Could not insert all of chunk into ",
-                to_string(kind),
-                " using options because of error.\n",
-                "\n",
-                "Chunk Size: ",
-                chunk |> length() |> to_string(),
-                "\n",
-                "Chunk:\n",
-                "\n",
-                inspect(chunk, limit: :infinity, printable_limit: :infinity),
-                "\n",
-                "\n",
-                "Options:\n",
-                "\n",
-                inspect(opts),
-                "\n",
-                "\n",
-                "Exception:\n",
-                "\n",
-                Exception.format(:error, exception, __STACKTRACE__)
-              ]
-            end)
-
-            Logger.configure(truncate: old_truncate)
-
-            # reraise to kill caller
-            reraise exception, __STACKTRACE__
-        end
-
-      if returning do
-        {count + total_count, acc ++ inserted}
-      else
-        {count + total_count, nil}
-      end
-    end)
-  end
-
-  def stream_in_transaction(query, fun) when is_function(fun, 1) do
-    transaction(
-      fn ->
-        query
-        |> stream(timeout: :infinity)
-        |> fun.()
-      end,
-      timeout: :infinity
-    )
-  end
-
-  def stream_each(query, fun) when is_function(fun, 1) do
-    stream_in_transaction(query, &Enum.each(&1, fun))
-  end
-
-  def stream_reduce(query, initial, reducer) when is_function(reducer, 2) do
-    stream_in_transaction(query, &Enum.reduce(&1, initial, reducer))
-  end
-
-
 end
